refactor(browser): move browser toolset out of core/ into toolsets/
Browser is a toolset (activated via ToolsetManager), not a core tool.
Move it to toolsets/browser.ts to match the convention used by
recordings and session-history toolsets.

- Remove createToolRuntime/truncationMiddleware usage from browser tools
- Declare truncation on the Toolset type so the manager applies it
- Add truncation field to Toolset interface
- Delete tools/core/browser.ts (merged into toolsets/browser.ts)
